Similar change for claim details
